Added basic not-empty validation to fields.

+++ client/src/components/Destroy.tsx
@@ -8,7 +8,7 @@ import { Button, Card, Col, Form } from 'react-bootstrap';
  * Component for destroying a key / value pair in Redis.
  * @returns JSX.Element
  */
-const Destroy = (): JSX.Element  => {
+const Destroy = (): JSX.Element => {
   // Set state related to controlled form inputs and API response
   const [key, setKey] = useState('');
   const [value, setValue] = useState('');
@@ -17,19 +17,22 @@ const Destroy = (): JSX.Element  => {
    * When the user clicks "Delete" this function is called.
    * @returns Void
    */
-   const onSubmit = (): void => {
-    // POST the API 
-    fetch(`http://localhost:8000/api/keys/${key}/`, {
-      method: 'DELETE'
-    }) 
-      .then(response => response.json())
-      .then(response => {
-        setValue(response['value']);
+  const onSubmit = (): void => {
+    // Only submit if field is populted
+    if (key) {
+      // POST the API
+      fetch(`http://localhost:8000/api/keys/${key}/`, {
+        method: 'DELETE',
       })
-      .catch(() => {
-        console.log('Error');
-      })
-  }
+        .then((response) => response.json())
+        .then((response) => {
+          setValue(response['value']);
+        })
+        .catch(() => {
+          console.log('Error');
+        });
+    }
+  };
 
   return (
     <Col>
@@ -37,7 +40,7 @@ const Destroy = (): JSX.Element  => {
         <Card.Header>Delete a Given Key</Card.Header>
         <Card.Body>
           <Card.Title>Instructions</Card.Title>
-          <Card.Text>Enter the Key and click Delete to delete a given key.</Card.Text>
+          <Card.Text>Enter the Key and click Delete to delete key.</Card.Text>
           <Form className='d-grid gap-2'>
             <Form.Group className='mb-3' controlId='formKey'>
               <Form.Label>Key</Form.Label>
